+++ src/core/temperaments/GOTemperament.cpp
@@ -10,10 +10,14 @@
 #include <wx/intl.h>
 
 GOTemperament::GOTemperament(wxString name, wxString group)
-  : m_Group(group), m_Name(name), m_Title(wxGetTranslation(name)) {}
+  : m_Group(group),
+    m_GroupTitle(wxEmptyString),
+    m_Name(name),
+    m_Title(wxEmptyString) {}
 
-GOTemperament::GOTemperament(wxString name, wxString title, wxString group)
-  : m_Group(group), m_Name(name), m_Title(title) {}
+GOTemperament::GOTemperament(
+  wxString name, wxString title, wxString group, wxString groupTitle)
+  : m_Group(group), m_GroupTitle(groupTitle), m_Name(name), m_Title(title) {}
 
 GOTemperament::~GOTemperament() {}
 
@@ -30,6 +34,12 @@ float GOTemperament::GetOffset(
 
 wxString GOTemperament::GetName() const { return m_Name; }
 
-wxString GOTemperament::GetTitle() const { return m_Title; }
+wxString GOTemperament::GetTitle() const {
+  return m_Title.IsEmpty() ? wxGetTranslation(m_Name) : m_Title;
+}
 
 wxString GOTemperament::GetGroup() const { return m_Group; }
+
+wxString GOTemperament::GetGroupTitle() const {
+  return m_GroupTitle.IsEmpty() ? wxGetTranslation(m_Group) : m_GroupTitle;
+}